+++ src/sst/selftests/refresh.py
@@ -1,0 +1,11 @@
+from sst.actions import *
+
+go_to('/')
+
+assert_title('The Page Title')
+assert_url('/')
+
+refresh()
+
+assert_title('The Page Title')
+assert_url('/')