+++ docs/changelog.rst
@@ -15,6 +15,7 @@ version **0.2.1** (2012 ??? ??)
 
 * abilty to handle `file:` based urls (static, non-http)
 * added `assert_equal` and `assert_not_equal` actions
+* added `refresh` action
 
 version **0.2.0** (2012 Feb 26)
 *******************************